+++ src/Facil.Runtime.CSharp/GeneratedCodeUtils.cs
@@ -18,49 +18,65 @@ namespace Facil.Runtime.CSharp
     [EditorBrowsable(EditorBrowsableState.Never)]
     public static class GeneratedCodeUtils
     {
-        private static async Task LoadTempTablesAsync(SqlConnection conn, IEnumerable<TempTableData> tempTableData, CancellationToken ct)
+        private static async Task LoadTempTablesAsync(SqlConnection conn, IEnumerable<TempTableData> tempTableData, SqlTransaction? tran, CancellationToken ct)
         {
             foreach (var data in tempTableData)
             {
-                using var cmd = conn.CreateCommand();
+                await using var cmd = conn.CreateCommand();
+                if (tran != null) cmd.Transaction = tran;
                 // Note: If there is ever a need for letting users configure the command,
                 // do not use the configureCmd parameter passed to methods on this class,
                 // which also sets any parameters.
                 cmd.CommandText = data.Definition;
                 await cmd.ExecuteNonQueryAsync(ct);
 
-                using var bulkCopy = new SqlBulkCopy(conn) { DestinationTableName = data.DestinationTableName };
+                using var bulkCopy = new SqlBulkCopy(conn, SqlBulkCopyOptions.Default, tran) { DestinationTableName = data.DestinationTableName };
                 data.ConfigureBulkCopy(bulkCopy);
                 var reader = new TempTableLoader(data.NumFields, data.Data);
                 await bulkCopy.WriteToServerAsync(reader, ct);
             }
         }
 
-        private static void LoadTempTables(SqlConnection conn, IEnumerable<TempTableData> tempTableData)
+        private static void LoadTempTables(SqlConnection conn, IEnumerable<TempTableData> tempTableData, SqlTransaction? tran)
         {
             foreach (var data in tempTableData)
             {
                 using var cmd = conn.CreateCommand();
+                if (tran != null) cmd.Transaction = tran;
                 // Note: If there is ever a need for letting users configure the command,
                 // do not use the configureCmd parameter passed to methods on this class,
                 // which also sets any parameters.
                 cmd.CommandText = data.Definition;
                 cmd.ExecuteNonQuery();
 
-                using var bulkCopy = new SqlBulkCopy(conn) { DestinationTableName = data.DestinationTableName };
+                using var bulkCopy = new SqlBulkCopy(conn, SqlBulkCopyOptions.Default, tran) { DestinationTableName = data.DestinationTableName };
                 data.ConfigureBulkCopy(bulkCopy);
                 var reader = new TempTableLoader(data.NumFields, data.Data);
                 bulkCopy.WriteToServer(reader);
             }
         }
 
-        public static async Task<List<T>> ExecuteQueryEagerAsync<T>(SqlConnection? existingConn, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData, CancellationToken ct)
+        private static void ConfigureCommand(SqlCommand cmd, SqlTransaction? tran, Action<SqlCommand> configureCmd)
+        {
+            configureCmd(cmd);
+            if (tran != null)
+            {
+                if (cmd.Transaction != null)
+                {
+                    throw new Exception("Transaction must not be set both WithConnection and ConfigureCommand; prefer WithConnection");
+                }
+
+                cmd.Transaction = tran;
+            }
+        }
+
+        public static async Task<List<T>> ExecuteQueryEagerAsync<T>(SqlConnection? existingConn, SqlTransaction? tran, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData, CancellationToken ct)
         {
             if (existingConn is not null)
             {
-                await LoadTempTablesAsync(existingConn, tempTableData, ct);
+                await LoadTempTablesAsync(existingConn, tempTableData, tran, ct);
                 using var cmd = existingConn.CreateCommand();
-                configureCmd(cmd);
+                ConfigureCommand(cmd, tran, configureCmd);
                 using var reader = await cmd.ExecuteReaderAsync(CommandBehavior.SingleResult, ct).ConfigureAwait(false);
                 var list = new List<T>();
                 if (reader.HasRows) {
@@ -77,7 +93,7 @@ namespace Facil.Runtime.CSharp
                 using var conn = new SqlConnection(connStr);
                 configureNewConn(conn);
                 await conn.OpenAsync(ct).ConfigureAwait(false);
-                await LoadTempTablesAsync(conn, tempTableData, ct);
+                await LoadTempTablesAsync(conn, tempTableData, tran, ct);
                 using var cmd = conn.CreateCommand();
                 configureCmd(cmd);
                 using var reader = await cmd.ExecuteReaderAsync(CommandBehavior.SingleResult, ct).ConfigureAwait(false);
@@ -98,13 +114,13 @@ namespace Facil.Runtime.CSharp
             }
         }
 
-        public static async Task<List<T>> ExecuteQueryEagerAsyncWithSyncRead<T>(SqlConnection? existingConn, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData, CancellationToken ct)
+        public static async Task<List<T>> ExecuteQueryEagerAsyncWithSyncRead<T>(SqlConnection? existingConn, SqlTransaction? tran, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData, CancellationToken ct)
         {
             if (existingConn is not null)
             {
-                await LoadTempTablesAsync(existingConn, tempTableData, ct);
+                await LoadTempTablesAsync(existingConn, tempTableData, tran, ct);
                 using var cmd = existingConn.CreateCommand();
-                configureCmd(cmd);
+                ConfigureCommand(cmd, tran, configureCmd);
                 using var reader = await cmd.ExecuteReaderAsync(CommandBehavior.SingleResult, ct).ConfigureAwait(false);
                 var list = new List<T>();
                 if (reader.HasRows)
@@ -122,7 +138,7 @@ namespace Facil.Runtime.CSharp
                 using var conn = new SqlConnection(connStr);
                 configureNewConn(conn);
                 await conn.OpenAsync(ct).ConfigureAwait(false);
-                await LoadTempTablesAsync(conn, tempTableData, ct);
+                await LoadTempTablesAsync(conn, tempTableData, tran, ct);
                 using var cmd = conn.CreateCommand();
                 configureCmd(cmd);
                 using var reader = await cmd.ExecuteReaderAsync(CommandBehavior.SingleResult, ct).ConfigureAwait(false);
@@ -143,13 +159,13 @@ namespace Facil.Runtime.CSharp
             }
         }
 
-        public static List<T> ExecuteQueryEager<T>(SqlConnection? existingConn, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData)
+        public static List<T> ExecuteQueryEager<T>(SqlConnection? existingConn, SqlTransaction? tran, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData)
         {
             if (existingConn is not null)
             {
-                LoadTempTables(existingConn, tempTableData);
+                LoadTempTables(existingConn, tempTableData, tran);
                 using var cmd = existingConn.CreateCommand();
-                configureCmd(cmd);
+                ConfigureCommand(cmd, tran, configureCmd);
                 using var reader = cmd.ExecuteReader(CommandBehavior.SingleResult);
                 var list = new List<T>();
                 if (reader.HasRows)
@@ -167,7 +183,7 @@ namespace Facil.Runtime.CSharp
                 using var conn = new SqlConnection(connStr);
                 configureNewConn(conn);
                 conn.Open();
-                LoadTempTables(conn, tempTableData);
+                LoadTempTables(conn, tempTableData, tran);
                 using var cmd = conn.CreateCommand();
                 configureCmd(cmd);
                 using var reader = cmd.ExecuteReader(CommandBehavior.SingleResult);
@@ -188,13 +204,13 @@ namespace Facil.Runtime.CSharp
             }
         }
 
-        public static async IAsyncEnumerable<T> ExecuteQueryLazyAsync<T>(SqlConnection? existingConn, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData, [EnumeratorCancellation] CancellationToken ct)
+        public static async IAsyncEnumerable<T> ExecuteQueryLazyAsync<T>(SqlConnection? existingConn, SqlTransaction? tran, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData, [EnumeratorCancellation] CancellationToken ct)
         {
             if (existingConn is not null)
             {
-                await LoadTempTablesAsync(existingConn, tempTableData, ct);
+                await LoadTempTablesAsync(existingConn, tempTableData, tran, ct);
                 using var cmd = existingConn.CreateCommand();
-                configureCmd(cmd);
+                ConfigureCommand(cmd, tran, configureCmd);
                 using var reader = await cmd.ExecuteReaderAsync(CommandBehavior.SingleResult, ct).ConfigureAwait(false);
                 if (reader.HasRows)
                 {
@@ -210,7 +226,7 @@ namespace Facil.Runtime.CSharp
                 using var conn = new SqlConnection(connStr);
                 configureNewConn(conn);
                 await conn.OpenAsync(ct).ConfigureAwait(false);
-                await LoadTempTablesAsync(conn, tempTableData, ct);
+                await LoadTempTablesAsync(conn, tempTableData, tran, ct);
                 using var cmd = conn.CreateCommand();
                 configureCmd(cmd);
                 using var reader = await cmd.ExecuteReaderAsync(CommandBehavior.SingleResult, ct).ConfigureAwait(false);
@@ -229,13 +245,13 @@ namespace Facil.Runtime.CSharp
             }
         }
 
-        public static async IAsyncEnumerable<T> ExecuteQueryLazyAsyncWithSyncRead<T>(SqlConnection? existingConn, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData, [EnumeratorCancellation] CancellationToken ct)
+        public static async IAsyncEnumerable<T> ExecuteQueryLazyAsyncWithSyncRead<T>(SqlConnection? existingConn, SqlTransaction? tran, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData, [EnumeratorCancellation] CancellationToken ct)
         {
             if (existingConn is not null)
             {
-                await LoadTempTablesAsync(existingConn, tempTableData, ct);
+                await LoadTempTablesAsync(existingConn, tempTableData, tran, ct);
                 using var cmd = existingConn.CreateCommand();
-                configureCmd(cmd);
+                ConfigureCommand(cmd, tran, configureCmd);
                 using var reader = await cmd.ExecuteReaderAsync(CommandBehavior.SingleResult, ct).ConfigureAwait(false);
                 if (reader.HasRows)
                 {
@@ -251,7 +267,7 @@ namespace Facil.Runtime.CSharp
                 using var conn = new SqlConnection(connStr);
                 configureNewConn(conn);
                 await conn.OpenAsync(ct).ConfigureAwait(false);
-                await LoadTempTablesAsync(conn, tempTableData, ct);
+                await LoadTempTablesAsync(conn, tempTableData, tran, ct);
                 using var cmd = conn.CreateCommand();
                 configureCmd(cmd);
                 using var reader = await cmd.ExecuteReaderAsync(CommandBehavior.SingleResult, ct).ConfigureAwait(false);
@@ -270,13 +286,13 @@ namespace Facil.Runtime.CSharp
             }
         }
 
-        public static IEnumerable<T> ExecuteQueryLazy<T>(SqlConnection? existingConn, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData)
+        public static IEnumerable<T> ExecuteQueryLazy<T>(SqlConnection? existingConn, SqlTransaction? tran, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData)
         {
             if (existingConn is not null)
             {
-                LoadTempTables(existingConn, tempTableData);
+                LoadTempTables(existingConn, tempTableData, tran);
                 using var cmd = existingConn.CreateCommand();
-                configureCmd(cmd);
+                ConfigureCommand(cmd, tran, configureCmd);
                 using var reader = cmd.ExecuteReader(CommandBehavior.SingleResult);
                 if (reader.HasRows)
                 {
@@ -292,7 +308,7 @@ namespace Facil.Runtime.CSharp
                 using var conn = new SqlConnection(connStr);
                 configureNewConn(conn);
                 conn.Open();
-                LoadTempTables(conn, tempTableData);
+                LoadTempTables(conn, tempTableData, tran);
                 using var cmd = conn.CreateCommand();
                 configureCmd(cmd);
                 using var reader = cmd.ExecuteReader(CommandBehavior.SingleResult);
@@ -311,13 +327,13 @@ namespace Facil.Runtime.CSharp
             }
         }
 
-        public static async Task<FSharpOption<T>> ExecuteQuerySingleAsync<T>(SqlConnection? existingConn, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData, CancellationToken ct)
+        public static async Task<FSharpOption<T>> ExecuteQuerySingleAsync<T>(SqlConnection? existingConn, SqlTransaction? tran, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData, CancellationToken ct)
         {
             if (existingConn is not null)
             {
-                await LoadTempTablesAsync(existingConn, tempTableData, ct);
+                await LoadTempTablesAsync(existingConn, tempTableData, tran, ct);
                 using var cmd = existingConn.CreateCommand();
-                configureCmd(cmd);
+                ConfigureCommand(cmd, tran, configureCmd);
                 using var reader = await cmd.ExecuteReaderAsync(CommandBehavior.SingleResult | CommandBehavior.SingleRow, ct).ConfigureAwait(false);
                 if (await reader.ReadAsync(ct).ConfigureAwait(false))
                 {
@@ -331,7 +347,7 @@ namespace Facil.Runtime.CSharp
                 using var conn = new SqlConnection(connStr);
                 configureNewConn(conn);
                 await conn.OpenAsync(ct).ConfigureAwait(false);
-                await LoadTempTablesAsync(conn, tempTableData, ct);
+                await LoadTempTablesAsync(conn, tempTableData, tran, ct);
                 using var cmd = conn.CreateCommand();
                 configureCmd(cmd);
                 using var reader = await cmd.ExecuteReaderAsync(CommandBehavior.SingleResult | CommandBehavior.SingleRow, ct).ConfigureAwait(false);
@@ -348,13 +364,13 @@ namespace Facil.Runtime.CSharp
             }
         }
 
-        public static async Task<FSharpValueOption<T>> ExecuteQuerySingleAsyncVoption<T>(SqlConnection? existingConn, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData, CancellationToken ct)
+        public static async Task<FSharpValueOption<T>> ExecuteQuerySingleAsyncVoption<T>(SqlConnection? existingConn, SqlTransaction? tran, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData, CancellationToken ct)
         {
             if (existingConn is not null)
             {
-                await LoadTempTablesAsync(existingConn, tempTableData, ct);
+                await LoadTempTablesAsync(existingConn, tempTableData, tran, ct);
                 using var cmd = existingConn.CreateCommand();
-                configureCmd(cmd);
+                ConfigureCommand(cmd, tran, configureCmd);
                 using var reader = await cmd.ExecuteReaderAsync(CommandBehavior.SingleResult | CommandBehavior.SingleRow, ct).ConfigureAwait(false);
                 if (await reader.ReadAsync(ct).ConfigureAwait(false))
                 {
@@ -368,7 +384,7 @@ namespace Facil.Runtime.CSharp
                 using var conn = new SqlConnection(connStr);
                 configureNewConn(conn);
                 await conn.OpenAsync(ct).ConfigureAwait(false);
-                await LoadTempTablesAsync(conn, tempTableData, ct);
+                await LoadTempTablesAsync(conn, tempTableData, tran, ct);
                 using var cmd = conn.CreateCommand();
                 configureCmd(cmd);
                 using var reader = await cmd.ExecuteReaderAsync(CommandBehavior.SingleResult | CommandBehavior.SingleRow, ct).ConfigureAwait(false);
@@ -385,13 +401,13 @@ namespace Facil.Runtime.CSharp
             }
         }
 
-        public static FSharpOption<T> ExecuteQuerySingle<T>(SqlConnection? existingConn, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData)
+        public static FSharpOption<T> ExecuteQuerySingle<T>(SqlConnection? existingConn, SqlTransaction? tran, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData)
         {
             if (existingConn is not null)
             {
-                LoadTempTables(existingConn, tempTableData);
+                LoadTempTables(existingConn, tempTableData, tran);
                 using var cmd = existingConn.CreateCommand();
-                configureCmd(cmd);
+                ConfigureCommand(cmd, tran, configureCmd);
                 using var reader = cmd.ExecuteReader(CommandBehavior.SingleResult | CommandBehavior.SingleRow);
                 if (reader.Read())
                 {
@@ -405,7 +421,7 @@ namespace Facil.Runtime.CSharp
                 using var conn = new SqlConnection(connStr);
                 configureNewConn(conn);
                 conn.Open();
-                LoadTempTables(conn, tempTableData);
+                LoadTempTables(conn, tempTableData, tran);
                 using var cmd = conn.CreateCommand();
                 configureCmd(cmd);
                 using var reader = cmd.ExecuteReader(CommandBehavior.SingleResult | CommandBehavior.SingleRow);
@@ -422,13 +438,13 @@ namespace Facil.Runtime.CSharp
             }
         }
 
-        public static FSharpValueOption<T> ExecuteQuerySingleVoption<T>(SqlConnection? existingConn, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData)
+        public static FSharpValueOption<T> ExecuteQuerySingleVoption<T>(SqlConnection? existingConn, SqlTransaction? tran, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, Action<SqlDataReader> initOrdinals, Func<SqlDataReader, T> getItem, IEnumerable<TempTableData> tempTableData)
         {
             if (existingConn is not null)
             {
-                LoadTempTables(existingConn, tempTableData);
+                LoadTempTables(existingConn, tempTableData, tran);
                 using var cmd = existingConn.CreateCommand();
-                configureCmd(cmd);
+                ConfigureCommand(cmd, tran, configureCmd);
                 using var reader = cmd.ExecuteReader(CommandBehavior.SingleResult | CommandBehavior.SingleRow);
                 if (reader.Read())
                 {
@@ -442,7 +458,7 @@ namespace Facil.Runtime.CSharp
                 using var conn = new SqlConnection(connStr);
                 configureNewConn(conn);
                 conn.Open();
-                LoadTempTables(conn, tempTableData);
+                LoadTempTables(conn, tempTableData, tran);
                 using var cmd = conn.CreateCommand();
                 configureCmd(cmd);
                 using var reader = cmd.ExecuteReader(CommandBehavior.SingleResult | CommandBehavior.SingleRow);
@@ -459,13 +475,13 @@ namespace Facil.Runtime.CSharp
             }
         }
 
-        public static async Task<int> ExecuteNonQueryAsync<T>(SqlConnection? existingConn, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, IEnumerable<TempTableData> tempTableData, CancellationToken ct)
+        public static async Task<int> ExecuteNonQueryAsync<T>(SqlConnection? existingConn, SqlTransaction? tran, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, IEnumerable<TempTableData> tempTableData, CancellationToken ct)
         {
             if (existingConn is not null)
             {
-                await LoadTempTablesAsync(existingConn, tempTableData, ct);
+                await LoadTempTablesAsync(existingConn, tempTableData, tran, ct);
                 using var cmd = existingConn.CreateCommand();
-                configureCmd(cmd);
+                ConfigureCommand(cmd, tran, configureCmd);
                 return await cmd.ExecuteNonQueryAsync(ct).ConfigureAwait(false);
             }
             else if (connStr is not null)
@@ -473,7 +489,7 @@ namespace Facil.Runtime.CSharp
                 using var conn = new SqlConnection(connStr);
                 configureNewConn(conn);
                 conn.Open();
-                await LoadTempTablesAsync(conn, tempTableData, ct);
+                await LoadTempTablesAsync(conn, tempTableData, tran, ct);
                 using var cmd = conn.CreateCommand();
                 configureCmd(cmd);
                 return await cmd.ExecuteNonQueryAsync(ct).ConfigureAwait(false);
@@ -484,13 +500,13 @@ namespace Facil.Runtime.CSharp
             }
         }
 
-        public static int ExecuteNonQuery<T>(SqlConnection? existingConn, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, IEnumerable<TempTableData> tempTableData)
+        public static int ExecuteNonQuery<T>(SqlConnection? existingConn, SqlTransaction? tran, string? connStr, Action<SqlConnection> configureNewConn, Action<SqlCommand> configureCmd, IEnumerable<TempTableData> tempTableData)
         {
             if (existingConn is not null)
             {
-                LoadTempTables(existingConn, tempTableData);
+                LoadTempTables(existingConn, tempTableData, tran);
                 using var cmd = existingConn.CreateCommand();
-                configureCmd(cmd);
+                ConfigureCommand(cmd, tran, configureCmd);
                 return cmd.ExecuteNonQuery();
             }
             else if (connStr is not null)
@@ -498,7 +514,7 @@ namespace Facil.Runtime.CSharp
                 using var conn = new SqlConnection(connStr);
                 configureNewConn(conn);
                 conn.Open();
-                LoadTempTables(conn, tempTableData);
+                LoadTempTables(conn, tempTableData, tran);
                 using var cmd = conn.CreateCommand();
                 configureCmd(cmd);
                 return cmd.ExecuteNonQuery();